+++ CHANGES.md
@@ -4,7 +4,9 @@ Full changelog
 v0.9.0 (unreleased)
 -----------------
 
-* No changes yet.
+* Implement new widget with better control over exporting to Plotly. The
+  preference pane for Plotly export has now been removed in favor of this new
+  way to set the options. [#1057]
 
 v0.8.3 (unreleased)
 -------------------